added separate playfield and buffer sizes for EM engine caves
The cave playfield size (or just cave size) is the size of the visible
playfield (potentially containing an additional steelwall border, if
the level was converted from R'n'D format and does not contain a
steelwall border).

The cave buffer size is the size of the internal cave buffer (which
also contains surrounding internal game engine elements) that is used
by the EM game engine logic.

Currently the native cave structure has the same size as the internal
cave buffer; this should be changed.