added support for EM wrap-around levels also to level editor
Previously, when running a level created with the R'n'D level editor
that uses the EM game engine, that level was forced to have a border
of indestructible game elements, which made it impossible to create
wrap-around levels to be played with the EM engine. This limitation
was removed, so no additional steel wall will ever be added around a
level.