removed test/debug code for EM graphics engine
This commit removes the testing and debugging capabilities for the
native Emerald Mine graphics engine (that can, but does not have to
be used with the native Emerald Mine game engine). The main purpose
of this test code was to check if the R'n'D "graphicsinfo.conf" style
graphics definitions, when used with the native Emerald Mine engine,
map exactly to this engine's hard-coded graphics definitions, to be
able to flexibly use R'n'D style graphics definitions also for the
native Emerald Mine game engine.

To test a given "graphicsinfo.conf" file against the internal graphics
definitions of the native Emerald Mine engine, its definitions have to
match an "objects" and "players" graphics file which is hard-coded in
the R'n'D graphics configuration file "conf_gfx.c" using the tokens
"emc_object" and "emc_sprite". (To prevent the game from using them in
non-debugging context, it normally refers to existing default artwork
instead of the test graphics files mentioned above.)

When testing or debugging, set these file names to "emc_objects.png"
and "emc_players.png" and enable upscaling to double size (assuming
the tiles in the test artwork have their original 16x16 pixel size).
Also set "DEBUG_EM_GFX" to "TRUE" in "src/tools.c". Then recompile.

When starting, the program performs the test, outputs all graphics
definitions that differ from the hard-coded internal definitions and
exits. To perform the test against the correct artwork, select the
artwork to test against before running the test. Usually the test is
executed using the graphics set "gfx_classic_emc.aga" from the Emerald
Mine Club level package, which uses the original artwork that is also
used by the game "Emerald Mine for X11".

(To perform the test described above, use a version of the code prior
to this commit.)