changed handling of chain explosions in EM engine
So far, chain explosions in the native EM game engine were not handled
like in the original Kingsoft or EMC player on the Amiga, which can be
tested with a burning dynamite surrounded by eight bugs, which explode
to only one diamond (and lots of emeralds) instead of four diamonds.

This was changed now (while keeping the old behaviour for old tapes).