extended maximum number of android clone elements from 16 to 32
This change was needed because some native EM/EMC levels (like level
71 from level set "EMC Mine 4") filled more than 16 clone elements
into the element list for the "android" game element, so it was not
possible to inspect all clone elements in the level editor (or build
a similar level in the editor). While the level ran just fine in the
native EM engine, the maximum number of android clone elements was
increased to 32 for the R'n'D level editor and level format. (Levels
using the previous chunk that was limited to 16 game elements can
still be loaded and played, but new levels will now be saved with the
new, extended chunk.)