fixed selecting ECS/AGA graphics if only ECS or AGA graphics exists
In rare cases (like in EMC level set "emc_no_one_mine_19"), there are
only ECS or AGA graphics available, but not both (as it is usually the
case). If this happened, the available graphics set was not selected
if the ECS/AGA preference settings did not match (for example, if ECS
graphics are preferred, but only AGA graphics are available). As no
level set specific graphics were selected at all in such a case, this
resulted in falling back to the R'n'D default graphics.

This bug was fixed by ignoring the ECS/AGA graphics preference in such
cases, but always using the only available graphics set instead.