fixed drawing old screens for one video frame when skipping fading
This fixes a problem that was introduced with the last commit: When
skipping fading effects (like when changing screens by pressing the
"Escape" key), the special bitmaps used for fading screens are not
updated to the current source or target screens, which can result in
displaying old screens (displayed before pressing "Escape") for one
video frame.

(For example, this happened when pressing "Escape" on a sub-screen
of the setup menu, then pressing "Escape" again on the setup main
screen, which then resulted in displaying the setup sub-screen in the
main menu for a single video frame, causing "garbage flickering".)

(This change effectively reverts commit 5eabbc4d. Negative effects
on cross-fading were not observed during tests, but may still exist.)