added being able to use CE condition "CE value/score gets 0"
While the custom element check "CE value/score gets 0 of [<element>]"
always worked fine, using the check "CE value/score gets 0" for the
same element never worked, because this required a previous change
action for changing the CE value or CE score in the same game frame,
but CEs can only change once in one game frame. (And checking in the
next game frame does not work either, because the CE value or score
does not change in the next frame, so it cannot be checked for zero.)

Resetting the CE's change counter after changing the CE value or score
makes it possible to test for zero directly afterwards.