added mouse click events to CE change events (experimental)
This addition is sort of fundamental to the R'n'D game engine in that
it extends the previous limitation of game engine input to only four
direction and two buttons, now also handling mouse clicks to elements
in the playfield as configurable custom element change events.

In contrast to the Mirror Magic game engine (which already handles
mouse input), recording mouse click events to tapes still has to be
implemented.