added new change action "move player new" to CE change actions
The difference to the already existing change action "move player" is
that "move player new" only moves the player if this starts a "new"
movement (that is, the player is either not moving at all yet, or has
just finished moving to the next tile).

This is useful to prevent the case of an apparent "two-step movement"
caused by moving the player right after a movement has just started,
which then adds a "second" movement after the "first" movement.