fixed element property that caused levels/tapes to be unsolvable
This fixes a bug introduced in version 2.0.1 that broke already
existing levels and tapes, and adds some compatibility code for levels
and tapes created later that require the previous behaviour.

This bugfix was needed for level set "rnd_sam_bateman", level 029,
for example, while the compatibility code was needed for level set
"rnd_gerhard_haeusler", level 023, for example.