fixed screen rotation for newer SDL versions on Android
Using the previous code with newer SDL versions on Android, screen
rotation did not work anymore, because the "resizable" flag seems
to be needed even when creating fullscreen surfaces to correctly
detect device rotation in recent SDL versions.

This was fixed by always using the "resizable" flag, regardless of
windowed or fullscreen surfaces on all platforms.