fixed screen size for newer SDL versions on Android
Using the previous code with newer SDL versions on Android, setting
the drawable screen size using the physical display size does not work
correctly anymore (as the display is usually larger than the drawable
screen area). This is corrected by using the effective size of the
renderer as soon as it is created.