fixed bug with reanimating killed player by CE condition
When checking for CE condition "explosion of <element>" using the
player as <element>, it is possible to create a new player right after
the player died (exploded).

However, this only worked when directly changing the CE to the player,
but not when using an extended change target (to create the player on
some adjacent tile next to the CE) that puts the new player on a tile
that is walkable (which is especially true for tiles that are empty).

This change fixes this bug. (The problematic code is in function
"CreateFieldExt()", but always checking for reanimating a killed
player in "InitPlayerField()" seems to be the better (more clean)
fix for this bug.)