fixed handling of pushed objects in EM engine for old tapes
This change fixes problems with emulating the behaviour of initially
pushed objects for old tapes. (With the new EM game engine, objects
that are pushed towards the player when initially created always kill
the player, while the object just stopped with the old engine.)

This is an addition to commit 1dd2190d.