fixed pushed objects going into acid in EM engine for old tapes
With the new EM game engine, objects that are pushed towards acid when
initially created fall into acid instead of replacing it (like in the
old engine). This change adds some compatibility code to use the old
behaviour when replaying old tapes.

This is an addition to commit f717d0c1.