fixed version number references for changes since version 4.2.0.0
Fixed some version number references to version 4.1.4.2 that was never
released as packages, but only via Git repository and that was nearly
identical to version 4.1.4.1. Especially it contained the same version
of the Emerald Mine game engine as version 4.1.4.1, so tapes recorded
with version 4.1.4.2 must be played with the old EM engine.

Starting with version 4.2.0.0, the new EM engine will be used.