added handling 64-bit random value bug with old tapes in EM engine
See commits 61c3da0 (incomplete fix) and c92edfe9 (final fix) for
details about this problem that was introduced (by a partial fix) in
version 3.3.1.0 and finally fixed in version 4.0.1.2, which resulted
in many tapes recorded between these versions with the EM game engine
to be broken.

To fix these broken tapes, it is required to set a new property bit in
all existing tape files that are affected by this bug.

The functionality to automatically patch these tapes will be added as
a new command line option.