added command to automatically test tapes and try to fix if broken
This adds a new command "autofix" that can be used to test and try to
fix existing tapes (by setting a property bit in the tape header).

The only patch currently available is "em_random_bug" that sets the
property bit for using the 64-bit random value workaround in the EM
game engine (see commit 1006b2db for details).

This command is an improved variant for the command "patch tapes"
introduced with commit 7c1d6252.

Thanks to Bela Lubkin for convincing me that this command is in fact a
good idea! :-)