added ability to undo (restore) the previously recorded tape
When the in-memory tape is erased with more than a minimum length of
playing time (currently set to 20 seconds), an internal copy of the
tape is created to be able to restore it. Just type ":undo-tape" or
":ut" to undo (restore) the last tape.

If the in-memory tape to be erased is shorter than the minimum length,
an undo tape is created from it if it is longer than the currently
stored undo tape.