fixed single-step mode for mouse click events
Before, when using mouse click events as CE change events, pressing a
mouse button in the playfield when in single-step mode was immediately
followed by entering pause mode again.

With this change, pause mode is not entered as long as a mouse button
is pressed. (This already works similarly in the MM game engine.)