added support for multiple user game data directories
This change adds support for game data directories for multiple users,
to store individual user configuration files (like setup data and last
played level sets and levels) and individual user content files (like
personal tape files, but not personal level sets and scores, which are
still stored to be accessible by all users).

Special treatment is done for the game data files of the first user,
which are still stored in the main game data directory like before,
so there are no changes in file and directory paths for the "default
user", regardless if multiple user support is enabled or disabled.