fixed bug with unnecessarily reloading custom artwork on ECS/AGA change
This bug occurs when changing the setup option for ECS/AGA preference
when there exists a level set with both ECS and AGA graphics sets in
the available tree of level sets, and results in (visually) attempting
to reload the current graphics even if the currently active level set
does not have any such graphics variants.

This fix prevents such unnecessary attempts to reload custom graphics,
and adds a better check for changed graphics (also needed if graphics
change from ECS to AGA variant) that really compares the previous with
the current graphics (which may result in the need to reload graphics).