fixed bug with text input gadgets when pressing "Escape" key
This bugs happens when incrementing or decrementing a numeric text
input gadget by 5 or 10 (using middle or right mouse button) so that
it gets a value beyond the allowed range. Although this will be
corrected, the "last value" variable still has the bad value, which
gets restored when clicking that text input gadget and then leaving
it by pressing the "Escape" key.