fixed bug with text area gadgets when pressing "Escape" key
This bugs happens when leaving a text area gadget by pressing the
"Escape" key. While this is allowed and does not cause the previous
text area content to be restored (as text area gadgets cannot be
left using the "Enter" key, and as there may be huge amount of text
being entered), the corresponding callback function is not called, so
the entered text is lost as soon as the gadget gets unmapped.