added asking to reset/remove user when clearing user name
When completely clearing a user name (in the main menu or on the name
selection screen, depending on the mutliple user setup setting), ask
to reset setup values for this user. This will cause the "setup.conf"
file to be deleted, and the name to be set back to "no name" (or the
default login name if that user is the first user in the user list).

When manually setting the name to "no name", ask to completely remove
that user from the game. This will cause the user directory (contains
settings, tapes etc.) to be renamed with a "renamed" suffix (together
with a timestamp), so it is not recognized by the program anymore, but
can still be copied back manually. This will only be done if the user
is not the first user in the user list.