improved inserting level sketch from clipboard to level editor
When pasting a level sketch to the level editor that has been copied
from the R'n'D forum, obsolete (historic) game elements will be mapped
to their current counterparts (just like it is done when loading an
old R'n'D level file). In addition, the corresponding game engine is
auto-detected (if possible) and set accordingly.