changed handling "conf" and "properties" buttons in level editor
When clicking "conf" and "properties" buttons in the level editor when
the previous screen is not the playfield screen, it can be confusing
how to return to the playfield screen (by clicking both buttons in the
"right" order). To make things easier, both buttons now return to the
playfield screen if the user is already on the corresponding screen.