fixed graphical bugs if digging or collecting caused player relocation
If digging or collecting (but not snapping!) certain tiles or elements
causes the player to be relocated ("teleported"), the corresponding
digging or collecting animation for this element will not be continued
anymore (because the player does not move towards this tile anymore).

To fix this problem, the digged or collected tile will be treated like
being snapped (using a temporary run-time element) for this edge case.
(The resulting engine behaviour is slightly different to the previous
behaviour, but should not break too many levels as this case should be
rather uncommon.)