changed calculating final game values when game is really over
Before this change, final game values (like final score and final
playing time or steps) were calculated before the game was really
over (but instead at the time the player started moving to the exit).

This especially caused calculating a wrong playing time that did not
include the time the player needed to finally reach the exit, which
was especially wrong when using steps instead of seconds/frames to
count the playing time, as the very last step was always missing in
this case.

This change fixes this bug by moving the code to calculate the final
game values to a later point where there is no player action anymore.