fixed bug with potentially changing global variable in thread
When automatically incrementing level after solved game, the global
variable to store the level number may be changed while the thread
to upload the score to the server is just using it, which may cause
the score to be submitted for the wrong level number.

To fix this, the (already existing) local variable with the level
number is used instead.