fixed triggering CE actions by digging or collecting
The improvement for CE actions by digging or collecting elements not
triggering before the dig or collect action was completely finished
had a very nasty bug in that such CE events when triggered by snapping
that element (which also counts as digging or collecting) just did not
work. This bug is fixed by this change.

This fixes commit ca2d6507.