fixed treating already handled click on gadget as "leave screen" event
This bug occured when clicking on the "previous level" or "next level"
gadget on the high score screen, followed by an empty scores screen,
which itself treats any button click as a "leave screen" event, thus
handling that mouse click twice, which results in that empty scores
screen to be immediately left (although it should wait for *another*
button click here, of course). This fix makes sure that button clicks
on gadgets that were already handled are not immediately handled again.