fixed crash bug for some graphics configurations with global border
This crash may happen if parts of a global screen border graphic are
configured to be drawn (like "border.draw_masked.DOOR: true") while
the global screen border graphic itself is configured to be undefined
(like "global.border.MAIN: [NONE]"), which causes an attempt to blit
from a bitmap that is NULL, resulting in a segmentation fault.

This commit adds a check to prevent this.