removed reloading artwork set for changed level set if not required
The code that was removed here did the following: Trigger reloading
artwork set after a different level set was selected if either the
previous or the current level set has custom artwork configured in
the file "levelinfo.conf". But this caused an attempt to reload new
artwork even if both the previous and current artwork are identical.

Now it is only checked if overriding artwork has changed, or if the
identifiers of the previous and current artwork have changed.