reverted commit 4ca952cbcd991a3e351b6292b37e47361b8f04d3
It turned out that the removed code did not only check for configured
artwork set in file "levelinfo.conf", but also for existing artwork
sub-directories in the previous and current level set when changing
level sets, causing such artwork to be ignored instead of activated.

This has to be handled differently to still work in such cases.