removed reloading artwork set for changed level set if not required (again)
The code that was changed here did the following: Trigger reloading
artwork set after a different level set was selected if either the
previous or the current level set has custom artwork configured in
the file "levelinfo.conf", or if either the previous or the current
level set contains a custom artwork sub-directory. But this caused an
attempt to reload new artwork even if both the previous and current
artwork in file "levelinfo.conf" are identical (but without custom
artwork sub-directories contained inside the level sets), which should
not happen.

Now it is only checked if overriding artwork has changed, or if the
identifiers of the previous and current artwork have changed, or if
either the previous or current level set contains a custom artwork
sub-directory, which then causes the artwork to be reloaded.

This was handled incorrectly by (reverted) commit 4ca952cb before.