fixed using individual program data path for Emscripten platform
Before, the Emscripten platform was treated as "Unix platform", which
should use ".rocksndiamonds" (or ".<program name>") as the data path.
As there is no "program name" on Emscripten platform, it is set to
"this" by the Emscripten environment, resulting in using ".this" for
all R'n'D based programs under the same domain in the IndexedDB store
of the browser, so there is no separation of individual program data
for different R'n'D based custom games, leading to shared data like
player name, setup settings and everything else, which is not desired.

This commit fixed this problem by using the program title instead of
the program's executable name (just like it is already done for the
Windows and Mac platforms.)