fixed compatibility handling for redefined old graphic "global.door"
If old graphic "global.door" is redefined, all individual definitions
accessing the same image file are changed to use the redefined image.

However, this did not work correctly for graphic definitions which are
cloned from other graphic definitions which are itself redefined and
do not use the same graphic as "global.door" anymore. In this case, do
not change these cloned graphics to use the redefined "global.door".