added showing score even if not added to local score file
This fixes a special case: If a score was not added to the local score
file (because all entries in the local score file have a better score,
and there are no free slots in the local score file anymore), the hall
of fame was completely skipped before. This was handled differently to
a score that was still not under the best 100 ranks on the high score
server for that level, but that was written to the local score file
(either because it was not full yet, or because it was better than the
last entry in the local score file).

This fix adds showing the score (with a rank of "???" on the last list
position, as usual) even if it was not written to the local score file.