fixed another problem with leaving scores screen on Android platform
On Android, when touching the screen anywhere in the hall of fame
after playing a game, this should either start the next game or
return to the main menu. However, this is done when touching the
screen with "finger down", not with "finger up again", sending an
additional "release" event in the main menu, triggering the menu
entry under the finger.

This fix triggers leaving the hall of fame on the "release" event
only.