fixed synchronization problem when configuring game controller
This fixes a bug that can happen due to bad initialization of a delay
counter (to limit the maximum time spent on handling events), causing
it to reach the maximum delay while the screen refresh is also waiting
for the video delay. Resetting the counter right after screen update
fixes this potential problem.