fixed synchronization problem when configuring keyboard
This is a somewhat similar bug like that in the previous commit, but
in this case only one event was processed for each screen update,
causing many events (like those produced from analog sticks that are
moved while the keyboard is about to be configured) to be processed
one after the other, each with a full video frame delay in between.

The fix works just like that used when configuring game controllers.