added level info config option to disable time limit for all levels
While it was already possible to disable the time limit in the setup
menu to play a level without any time limit, the resulting tapes only
replay successfully if the time limit is still disabled, of course.
Obviously, this is not the case on the central high score server.

However, the level sets in the "R'n'D jue 2022" games collection are
all designed to be played without time limit, and the time limit is
disabled in the default setup options of all games. This currently
has the following two disadvantages:

1. The resulting tapes fail on the high score server, which has time
   limit always enabled.
2. If the user of "R'n'D jur 2022" games enables the time limit in the
   setup menu, the levels cannot be solved anymore.

This patch adds a new time limit option to the level info config file,
which makes it possible to flag whole level sets as "to be played with
time limit disabled", regardless of what the individual user's setting
for "time limit" was set to.

This patch makes it possible to fix case (1) above by adding the new
level info config option to all affected levels on the high score
server, but would require a program update to also fix case (2) on the
client side.