added rule to Makefile to ensure reproducible builds
Declare build dependency to ensure consistent make execution order.
See https://reproducible-builds.org/ for why this is good.

Without this patch, parallelism would make a difference in
the produced config.o

This patch was done while working on reproducible builds for openSUSE.

Thanks to Bernhard M. Wiedemann for this patch!