added optional main menu button to directly show level set info
This is a special button for the main menu that will only be displayed
if files with level set information exist (like a "README" file in the
current level set directory).

The button will currently only be displayed if a valid screen position
is defined for it in the currently active graphics configuration.