changed "README" file for Niko Böhm's Tutorial level set
This changes the old ISO-8859-1 encoding to UTF-8 and adds using text
in different fonts and colors. The original "README" file is stored as
"README.orig" for reference, and a new, alternative version is stored
as "README.txt", which is not used by the game, but shows alternative
formatting possibilities using Markdown-style line breaks (using two
spaces at the end of lines) and pure text content without font usage.