added using GIC flags "request.button.player_X.draw_player"
This "graphicsinfo.conf" flag was completely ignored until now. It can
be used to control if player graphics should be drawn on top of the
four request buttons to select the network player in the setup menu,
or if only the buttons (without additional player graphics) should be
used (which only makes sense if custom player request buttons are used
with individual player graphics). The default value is "true".